backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
boto3>=1.34.0
python-multipart>=0.0.6
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
email-validator>=2.0.0
openai>=1.98.0
openai-agents==0.2.4
redis>=5.0.0
mailgun>=1.1.0